new file:   base/ParamList.h 	modified:   base/Timestamp.h 	modified:   log/LogFile.cpp 	modified:   log/LogStream.cpp 	modified:   log/Logger.cpp 	modified:   log/Logger.h 	new file:   project/Log/Log.cpp 	new file:   project/Log/Log.vcproj 	new file:   project/Log/ReadMe.txt 	new file:   project/Log/stdafx.cpp 	new file:   project/Log/stdafx.h 	new file:   project/TestLog/ReadMe.txt 	renamed:    test/TestLog/TestLog.cpp -> project/TestLog/TestLog.cpp 	new file:   project/TestLog/TestLog.vcproj 	new file:   project/TestLog/stdafx.cpp 	new file:   project/TestLog/stdafx.h 	new file:   project/TheFoxFramework.sln 	deleted:    test/TestLog/ReadMe.txt 	deleted:    test/TestLog/TestLog.sln 	deleted:    test/TestLog/TestLog.vcxproj 	deleted:    test/TestLog/stdafx.cpp 	deleted:    test/TestLog/stdafx.h 	deleted:    test/TestLog/targetver.h

+++ log/LogFile.cpp
@@ -227,7 +227,7 @@ void LogFile::makePath(String &dir)
     }
     _mkdir(filePath);
     
-    int pathLen = strlen(filePath);
+    size_t pathLen = strlen(filePath);
     if ('/' != filePath[pathLen - 1])
     {
         strcat(filePath, "/");
